Only call fs.realpathSync() if the file exists
Files which are deleted that Atom currently has open will have a path but won’t actually exist in the file system. This caused us to throw a TypeError (see #168). To avoid this, we simply use the path returned by TextEditor::getPath() if the file does not exist.

To avoid confusion, I have also renamed the `relativeActivePanePath` and `activePanePath` variables to a single `activePanePath` variable.

Fixes #168 and closes #174.