Only observe text editors
When opening a file in a pending tab from the tree view, the active pane will be the tree view (instead of the text editor). This causes us to throw the “Something went wrong internally” warning if there haven’t been any other active text editors (because the `text-editor` cache is still empty at that point). To avoid this, we use observeTextEditors() to only observe text editors (instead of every active pane change).

This further allows us to remove __checkIfTextEditor() and the initial storeSettings() call on startup (since observeTextEditors() will call back anyway). However, it requires us to pass the text editor to checkStyleSettings() since it used getActiveTextEditor() on its own.

Fixes #179 and closes #190.